overlays: Update display GPIO declarations
The 5.4 kernel changes the way a number of display drivers use GPIOs.
That change has exposed flaws/broken a number of display overlays, so
after a trickle of single-display patches this is a best-guess attempt
to fix the remainder. As none of these changes have been tested on real
displays there is a possibility that this either doesn't fix the
problem or even breaks something that was working - apologies if that
is the case.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>